05_Web_Development_&_APIs/AI Job Recommendation System(Web App)/requirements.txt
asgiref==3.11.0
Django==5.0.6
djangorestframework==3.16.1
psycopg2-binary==2.9.11
python-dotenv==1.2.1
sqlparse==0.5.4
Pillow>=10.0
sentence-transformers
torch
numpy
scikit-learn
django-crontab
pdfplumber
python-docx
gunicorn
dj-database-url
requests
whitenoise